requirements.txt
asgiref==3.11.1
boto3==1.42.46
botocore==1.42.46
Django==6.0.2
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
jmespath==1.1.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
